Add files via upload
In this project, we make a Restaurant Management System where we will explore the Tkinter library and its functions.

Tkinter is the standard GUI library for Python. Python when combined with Tkinter provides a fast and easy way to create GUI applications.

The objective of this project was to manage the restaurant more effectively and efficiently by computerizing meal ordering and billing.

 In this we have a versatile menu as well as a price for each item. The user enters the number of portions of order, and the backend calculates the total amount, including service fees and tax.

There are classes in the system’s backend that can perform functions including tax calculation, service charge calculation, and overall price of the meal including all taxes.